% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{add_derived_date_variables}
\alias{add_derived_date_variables}
\title{Add derived data variables}
\usage{
add_derived_date_variables(data)
}
\arguments{
\item{data}{Applications dataset}
}
\value{
Data with added date variables
}
\description{
Adds derived date-related variables to the applications data for temporal analysis.
}
