% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{add_derived_postcode_variables}
\alias{add_derived_postcode_variables}
\title{Add Postal Code info}
\usage{
add_derived_postcode_variables(data)
}
\arguments{
\item{data}{Dataset with postcodecijfers column}
}
\value{
Dataset with valid_postcode flag added
}
\description{
Adds numeric variable and validation flag for correct postcode to dataset.
}
