% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{add_derived_status_variables}
\alias{add_derived_status_variables}
\title{Add derived status variables}
\usage{
add_derived_status_variables(data)
}
\arguments{
\item{data}{Applications dataset}
}
\value{
Data with added status variables
}
\description{
Adds derived status-related variables to the applications data for analysis.
}
