% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{add_external_program_variables}
\alias{add_external_program_variables}
\title{Add Program information}
\usage{
add_external_program_variables(applications)
}
\arguments{
\item{applications}{Applications data tibble}
}
\value{
A tibble with enriched applications data
}
\description{
Enriches applications data with reference data (BRIN and programmes).
}
