% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{calculate_summary_stats}
\alias{calculate_summary_stats}
\title{Calculate Summary Statistics}
\usage{
calculate_summary_stats(data, grouping_level = "full")
}
\arguments{
\item{data}{Applications dataset}

\item{grouping_level}{String: level of analysis}
}
\value{
List with summary statistics
}
\description{
Calculates summary statistics for comparison purposes (averages, medians, etc.)
}
