% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{create_market_share_map}
\alias{create_market_share_map}
\title{Create geographic postcode distribution visualization}
\usage{
create_market_share_map(
  postcode_data,
  geo_data_path,
  target_year,
  institution_name
)
}
\arguments{
\item{postcode_data}{Postcode distribution data from calculate_market_share_by_postcode}

\item{geo_data_path}{Path to geographic data file (GeoPackage format)}

\item{target_year}{Year to visualize}

\item{institution_name}{Name of institution for labels}
}
\value{
Leaflet map object
}
\description{
Create geographic postcode distribution visualization
}
