% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_analysis.R
\name{filter_applications_on_brin}
\alias{filter_applications_on_brin}
\title{Filter Applications Data}
\usage{
filter_applications_on_brin(data, brin_codes = NULL)
}
\arguments{
\item{data}{Applications dataset (output from load_and_enrich_applications)}

\item{brin_codes}{Optional vector of BRIN codes to filter on}
}
\value{
Filtered tibble
}
\description{
Filters applications data based on specified criteria like school years and BRIN codes.
}
