% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{get_all_brin_codes}
\alias{get_all_brin_codes}
\title{Get All BRIN Codes from Data}
\usage{
get_all_brin_codes(data = NULL, min_applications = 50)
}
\arguments{
\item{data}{Applications dataset (optional, will load if not provided)}

\item{min_applications}{Minimum number of applications required (default: 50)}
}
\value{
Vector of BRIN codes
}
\description{
Extracts all unique BRIN codes from the applications data for iteration.
}
