% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_institutions}
\alias{get_institutions}
\title{Get Institutions}
\usage{
get_institutions()
}
\value{
Data frame with brin_code and school_name columns
}
\description{
Extracts available institutions from the data with BRIN codes and school names
}
