% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{load_applications}
\alias{load_applications}
\title{Load Applications Data}
\usage{
load_applications()
}
\value{
A tibble with applications data
}
\description{
Loads raw applications data from configured data source.
}
