% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{load_brin_info}
\alias{load_brin_info}
\title{Load BRIN Information}
\usage{
load_brin_info()
}
\value{
A tibble with BRIN information
}
\description{
Loads institutional information from BRIN reference data.
}
