% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{load_programmes_info}
\alias{load_programmes_info}
\title{Load Programmes Information}
\usage{
load_programmes_info()
}
\value{
A tibble with programmes information
}
\description{
Loads programme information from reference data.
}
