% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{open_config}
\alias{open_config}
\title{Open Configuration File}
\usage{
open_config()
}
\description{
Opens the project configuration file in RStudio editor if available,
otherwise prints the config path. Uses the config package's default
behavior to find the config file.
}
