% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_institution_report}
\alias{render_institution_report}
\title{Render Single Institution Report}
\usage{
render_institution_report(
  brin_code,
  school_name,
  year = 2024,
  suffix = "",
  input_file_path = "analysis/instelling_analysis.qmd",
  output_base_dir_relative = "../output"
)
}
\arguments{
\item{brin_code}{BRIN code of the institution}

\item{school_name}{Name of the school}

\item{year}{Numeric value of year to analyze}

\item{suffix}{Optional suffix for the output filename}

\item{input_file_path}{File path to the Quarto input file}

\item{output_base_dir_relative}{directory for output files, relative to the input path}
}
\value{
TRUE if rendering succeeded, FALSE otherwise
}
\description{
Render Single Institution Report
}
