% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{render_multiple_institution_reports}
\alias{render_multiple_institution_reports}
\title{Main Rendering Function}
\usage{
render_multiple_institution_reports(
  mode = "all",
  brin_codes = NULL,
  year = 2024
)
}
\arguments{
\item{mode}{Character: "specific" or "all"}

\item{brin_codes}{Vector of specific BRIN codes (when mode = "specific")}

\item{year}{Numeric value of year to analyze}
}
\description{
Main Rendering Function
}
